add doc comments, modularize, and clean up code.
