+++ src/Router/index.tsx
@@ -1,125 +1,15 @@
 import React from 'react';
-import { pick, resolve, match, createRoute } from '@/Router/helpers';
-import { globalHistory, navigate, createHistory, createMemorySource } from '@/Router/history';
-import { isRedirect } from '@/Router/Redirect';
+import { pick, resolve, match, createRoute, createNamedContext } from '@/Router/helpers';
+import { navigate, createHistory, createMemorySource } from '@/Router/history';
 
+/* imports to export at the top-level */
+import { Redirect } from '@/Router/Redirect';
+import { Location, ServerLocation, LocationProvider } from '@/Router/Location';
 import Link from '@/Router/Link';
 
-const createNamedContext = (name, defaultValue) => {
-  const Ctx = React.createContext(defaultValue);
-  Ctx.Consumer.displayName = `${name}.Consumer`;
-  Ctx.Provider.displayName = `${name}.Provider`;
-
-  return Ctx;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-// Location Context/Provider
-const LocationContext = createNamedContext('Location');
-
-// sets up a listener if there isn't one already so apps don't need to be
-// wrapped in some top level provider
-const Location = ({ children }) => (
-  <LocationContext.Consumer>
-    {context => (context ? children(context) : <LocationProvider>{children}</LocationProvider>)}
-  </LocationContext.Consumer>
-);
-
-interface PLocationProvider {
-  history?: any;
-  children?: (any) => JSX.Element;
-}
-interface SLocationProvider {
-  context: {
-    navigate: any;
-    location: any;
-  };
-  refs: {
-    unlisten: any;
-  };
-}
-
-class LocationProvider extends React.Component<PLocationProvider, SLocationProvider> {
-  public static defaultProps: Partial<PLocationProvider> = {
-    history: globalHistory
-  };
-
-  state = {
-    context: this.getContext(),
-    refs: { unlisten: null }
-  };
-
-  unmounted = null;
-
-  componentDidMount() {
-    this.state.refs.unlisten = this.props.history.listen(() => {
-      Promise.resolve().then(() => {
-        if (!this.unmounted) {
-          this.setState(() => ({ context: this.getContext() }));
-        }
-      });
-    });
-  }
-
-  componentDidCatch(error, info) {
-    if (isRedirect(error)) {
-      this.props.history.navigate(error.uri, { replace: true });
-    } else {
-      throw error;
-    }
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (prevState.context.location !== this.state.context.location) {
-      this.props.history._onTransitionComplete();
-    }
-  }
-
-  componentWillUnmount() {
-    this.unmounted = true;
-    this.state.refs.unlisten();
-  }
-
-  getContext() {
-    const {
-      props: {
-        history: { navigate, location }
-      }
-    } = this;
-
-    return { navigate, location };
-  }
-
-  render() {
-    const isChildFn = typeof this.props.children === 'function';
-
-    return (
-      <LocationContext.Provider value={this.state.context}>
-        {isChildFn ? this.props.children(this.state.context) : this.props.children || null}
-      </LocationContext.Provider>
-    );
-  }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-const ServerLocation = ({ url, children }) => (
-  <LocationContext.Provider
-    value={{
-      location: { pathname: url },
-      navigate: () => {
-        throw new Error("You can't call navigate on the server.");
-      }
-    }}
-  >
-    {children}
-  </LocationContext.Provider>
-);
-
-////////////////////////////////////////////////////////////////////////////////
 // Sets baseuri and basepath for nested routers and links
 const BaseContext = createNamedContext('Base', { baseuri: '/', basepath: '/' });
 
-////////////////////////////////////////////////////////////////////////////////
 /**
  * @description Main Router component that connects the matched Component to
  *              the contexts.
@@ -325,6 +215,7 @@ class FocusHandlerImpl extends React.Component<PFocusHandlerImpl, SFocusHandlerI
       location,
       ...domProps
     } = this.props;
+
     return (
       <Comp
         style={{ outline: 'none', ...style }}
@@ -366,7 +257,6 @@ const Match = ({ path, children }) => (
   </BaseContext.Consumer>
 );
 
-////////////////////////////////////////////////////////////////////////
 export {
   Location,
   LocationProvider,
@@ -378,5 +268,6 @@ export {
   navigate,
   BaseContext,
   /* Components */
-  Link
+  Link,
+  Redirect
 };
